feat(refactoring): Rename package to douki
