+++ .github/ISSUE_TEMPLATE/feature-request.yml
@@ -7,7 +7,7 @@ body:
   - type: markdown
     attributes:
       value: >
-        **Thank you for wanting to suggest a feature for Doxs!**
+        **Thank you for wanting to suggest a feature for Douki!**
 
 
         ⚠
@@ -17,7 +17,7 @@ body:
         already be implemented in a development release.
 
 
-        [issue search]: https://github.com/xmnlab/doxs.git/issues?q=is%3Aopen+is%3Aissue+label%3Afeat%2Fenhancement
+        [issue search]: https://github.com/xmnlab/douki.git/issues?q=is%3Aopen+is%3Aissue+label%3Afeat%2Fenhancement
 
   - type: textarea
     attributes:
@@ -28,7 +28,7 @@ body:
 
         What's the problem this feature will solve?
         What are you trying to do, that you are unable to achieve
-        with the **latest** release of Doxs?
+        with the **latest** release of Douki?
 
 
         * Provide examples of real-world use cases that this would enable
@@ -44,7 +44,7 @@ body:
         If there are external references or other GitHub Issues that are related
         please link them here as well.
       placeholder: >-
-        I am trying to do X with Doxs version x.y.z and I think that implementing
+        I am trying to do X with Douki version x.y.z and I think that implementing
         new feature Y would be very helpful for me and every other user because of Z.
     validations:
       required: true
@@ -70,9 +70,9 @@ body:
     attributes:
       label: Code of Conduct
       description: |
-        Read the [`Doxs` Code of Conduct][CoC] first.
+        Read the [`Douki` Code of Conduct][CoC] first.
 
-        [CoC]: https://github.com/xmnlab/doxs.git/coc/CODE_OF_CONDUCT.md
+        [CoC]: https://github.com/xmnlab/douki.git/coc/CODE_OF_CONDUCT.md
       options:
         - label: I agree to follow the Code of Conduct
           required: true
