+++ .github/workflows/release.yaml
@@ -30,7 +30,7 @@ jobs:
           miniforge-version: latest
           environment-file: conda/dev.yaml
           channels: conda-forge,nodefaults
-          activate-environment: doxs
+          activate-environment: douki
           auto-update-conda: true
           conda-solver: libmamba
