+++ .pre-commit-config.yaml
@@ -49,7 +49,7 @@ repos:
         name: mypy
         entry: mypy
         language: system
-        files: "src/doxs"
+        files: "src/douki"
         pass_filenames: true
         types:
           - python
@@ -67,7 +67,7 @@ repos:
         name: vulture
         entry: vulture
         language: system
-        files: "src/doxs"
+        files: "src/douki"
         description: Find unused Python code.
         pass_filenames: true
         types:
@@ -77,7 +77,7 @@ repos:
         name: mccabe
         entry: python -m mccabe --min 10
         language: system
-        files: "src/doxs"
+        files: "src/douki"
         pass_filenames: true
         types:
           - python
