+++ .releaserc.json
@@ -8,12 +8,12 @@
       {
         "replacements": [
           {
-            "files": ["src/doxs/__init__.py"],
+            "files": ["src/douki/__init__.py"],
             "from": "return '.*'  # semantic-release",
             "to": "return '${nextRelease.version}'  # semantic-release",
             "results": [
               {
-                "file": "src/doxs/__init__.py",
+                "file": "src/douki/__init__.py",
                 "hasChanged": true,
                 "numMatches": 1,
                 "numReplacements": 1
@@ -65,7 +65,7 @@
         "assets": [
           "pyproject.toml",
           "docs/changelog.md",
-          "src/doxs/__init__.py"
+          "src/douki/__init__.py"
         ],
         "message": "chore(release): ${nextRelease.version}"
       }
