+++ docs/changelog.md
@@ -1,36 +1,36 @@
 # Release Notes
 ---
 
-# [0.5.0](https://github.com/osl-incubator/doxs/compare/0.4.0...0.5.0) (2025-08-01)
+# [0.5.0](https://github.com/osl-incubator/douki/compare/0.4.0...0.5.0) (2025-08-01)
 
 
 ### Features
 
-* Improve the class method docstring generation ([#8](https://github.com/osl-incubator/doxs/issues/8)) ([ba6bf89](https://github.com/osl-incubator/doxs/commit/ba6bf89c05c0e837ca5509936dde1a3d8af969fd))
+* Improve the class method docstring generation ([#8](https://github.com/osl-incubator/douki/issues/8)) ([ba6bf89](https://github.com/osl-incubator/douki/commit/ba6bf89c05c0e837ca5509936dde1a3d8af969fd))
 
-# [0.4.0](https://github.com/osl-incubator/doxs/compare/0.3.0...0.4.0) (2025-08-01)
-
-
-### Features
-
-* Add initial schema validation ([#7](https://github.com/osl-incubator/doxs/issues/7)) ([a647142](https://github.com/osl-incubator/doxs/commit/a64714264bec158fdbb8395a66532fbcf8c6f458))
-
-# [0.3.0](https://github.com/osl-incubator/doxs/compare/0.2.0...0.3.0) (2025-08-01)
+# [0.4.0](https://github.com/osl-incubator/douki/compare/0.3.0...0.4.0) (2025-08-01)
 
 
 ### Features
 
-* Replace homemade Annotation by typing.Annotated ([#6](https://github.com/osl-incubator/doxs/issues/6)) ([f098bab](https://github.com/osl-incubator/doxs/commit/f098bab285039104e33618844058e2f571534a13))
+* Add initial schema validation ([#7](https://github.com/osl-incubator/douki/issues/7)) ([a647142](https://github.com/osl-incubator/douki/commit/a64714264bec158fdbb8395a66532fbcf8c6f458))
 
-# [0.2.0](https://github.com/osl-incubator/doxs/compare/0.1.0...0.2.0) (2025-07-31)
+# [0.3.0](https://github.com/osl-incubator/douki/compare/0.2.0...0.3.0) (2025-08-01)
+
+
+### Features
+
+* Replace homemade Annotation by typing.Annotated ([#6](https://github.com/osl-incubator/douki/issues/6)) ([f098bab](https://github.com/osl-incubator/douki/commit/f098bab285039104e33618844058e2f571534a13))
+
+# [0.2.0](https://github.com/osl-incubator/douki/compare/0.1.0...0.2.0) (2025-07-31)
 
 
 ### Bug Fixes
 
-* Add missing changes ([#2](https://github.com/osl-incubator/doxs/issues/2)) ([8a10850](https://github.com/osl-incubator/doxs/commit/8a108509297a4ff4488529c96012875e975c2a8a))
-* Fix semantic release configuration ([#5](https://github.com/osl-incubator/doxs/issues/5)) ([c25eeef](https://github.com/osl-incubator/doxs/commit/c25eeefbb94b4931f1bcd99cd9ddb2c97919e559))
+* Add missing changes ([#2](https://github.com/osl-incubator/douki/issues/2)) ([8a10850](https://github.com/osl-incubator/douki/commit/8a108509297a4ff4488529c96012875e975c2a8a))
+* Fix semantic release configuration ([#5](https://github.com/osl-incubator/douki/issues/5)) ([c25eeef](https://github.com/osl-incubator/douki/commit/c25eeefbb94b4931f1bcd99cd9ddb2c97919e559))
 
 
 ### Features
 
-* Add support for docstring in yaml format ([#4](https://github.com/osl-incubator/doxs/issues/4)) ([d9cbd9a](https://github.com/osl-incubator/doxs/commit/d9cbd9a731975c82e7842c8f03bbf1373a269ba3))
+* Add support for docstring in yaml format ([#4](https://github.com/osl-incubator/douki/issues/4)) ([d9cbd9a](https://github.com/osl-incubator/douki/commit/d9cbd9a731975c82e7842c8f03bbf1373a269ba3))
